feat: adds basic logic based on step, with no repeating

+++ src/specs/game/game.rs
@@ -58,11 +58,11 @@ pub enum GameSpecError {
 
     /// Some position has dimensions different than the board.
     #[error("Direction has invalid dimensions: {0:?}")]
-    InvalidDirectionDimensions(Vec<i8>),
+    InvalidDirectionDimensions(Vec<i16>),
 
     /// Some direction has values that are neither 1 nor -1.
     #[error("Direction has a value different than 1 or -1: {0:?}")]
-    InvalidDirectionValue(i8),
+    InvalidDirectionValue(i16),
 
     /// A specified position has been marked as disabled on the board.
     #[error("The specified position is disabled on the board: {0:?}")]
